Restructure KAPy package to improve loading speed
