+++ KAPy/__init__.py
@@ -1,9 +1,8 @@
 #Top level init.py for KAPy package
 
-from .configs import *
+from .helpers import *
+from .download import *
 from .primVars import *
 from .ensembles import *
-from .inputs import *
-from .helpers import *
 from .indicators import *
 from .workflow import *
